+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.3.0] - 2024-08-02
 ### Changed
 - **BREAKING**: remove `BorrowedEntry::into_owned()`.
 - Update `sdd` to v2.
@@ -45,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feuer Frei!
 
 <!-- next-url -->
-[Unreleased]: https://github.com/loyd/idr-ebr/compare/v0.2.3...HEAD
+[Unreleased]: https://github.com/loyd/idr-ebr/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/loyd/idr-ebr/compare/v0.2.3...v0.3.0
 [0.2.3]: https://github.com/loyd/idr-ebr/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/loyd/idr-ebr/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/loyd/idr-ebr/compare/v0.2.0...v0.2.1